+++ src/core/handlers/base.ts
@@ -1,7 +1,6 @@
 import { ICuiLogger, IUIInteractionProvider, CuiContext, ICuiComponentHandler, ICuiParsable, ICuiOpenable, ICuiClosable, CuiDevelopmentStateType, KeyDownEvent } from "../models/interfaces";
 import { CuiLoggerFactory } from "../factories/logger";
 import { CuiUtils } from "../models/utils";
-
 import { ICuiComponentMutationObserver, CuiComponentMutationHandler } from "../observers/mutations";
 import { AriaAttributes } from "../utils/aria";
 import { CuiActionsHelper } from "../helpers/helpers";