+++ src/core/models/events.ts
@@ -1,19 +1,26 @@
 import { ICuiComponentAction } from "../utils/actions";
+import { CuiHTMLElement } from "./interfaces";
 
-export interface CuiTargetChangeEvent {
+export interface EventBase {
+    name: string;
+    timestamp: number;
+    source: CuiHTMLElement;
+}
+
+export interface CuiTargetChangeEvent extends EventBase {
     top: number;
     left: number;
     target: Element;
     timestamp: number;
 }
 
-export interface CuiToggleEvent {
+export interface CuiToggleEvent extends EventBase {
     action: ICuiComponentAction;
     target: Element;
     timestamp: number;
 }
 
-export interface CuiScrollByEvent {
+export interface CuiScrollByEvent extends EventBase {
     to: number;
     by: number;
     target: Element;