+++ src/core/models/interfaces.ts
@@ -37,6 +37,10 @@ export interface ICuiComponentHandler {
     destroy(): void;
 }
 
+export interface ICuiElementHandlers {
+    [id: string]: ICuiComponentHandler;
+}
+
 export interface ICuiOpenable {
     open(args?: any): Promise<boolean>;
 }
@@ -91,9 +95,7 @@ export interface ICuiEventBus {
     detachByCuid(event: string, cuid: string): void;
 }
 
-export interface ICuiCallbackExecutor {
-    execute(callback: any, args: any[]): Promise<void>;
-}
+
 
 export interface CuiEventObj {
     callback: any;
@@ -104,9 +106,7 @@ export interface CuiEventReceiver {
     [id: string]: CuiEventObj;
 }
 
-export interface ICuiEventEmitHandler {
-    handle(receiver: CuiEventReceiver, cuid: string | null, args: any[]): Promise<void>;
-}
+
 
 export interface CuiContext {
     getId(): string;
@@ -150,6 +150,10 @@ export interface CuiHandlers {
 export interface CuiElement {
     $cuid: string | null;
     $handlers?: CuiHandlers;
+}
+
+export interface CuiHTMLElement extends HTMLElement, CuiElement {
+
 }
 
 export interface ICuiObserver {