+++ src/index.ts
@@ -1,6 +1,6 @@
 import { CuiInit } from './app/init';
 
-export const CUI_LIGHT_VERSION = "0.3.6";
+export const CUI_LIGHT_VERSION = "0.3.7";
 
 declare global {
     interface Window {