+++ changelog.md
@@ -1,5 +1,8 @@
 # bpd cUI Light changelog
 
+# [0.3.7] 2021-02-20
+* [Changed] cUI element creation process
+* [Fixed] Cleanup in Event bus
 # [0.3.6] 2021-02-19
 * [Added] Focus plugin to add proper class to body depending on input type - ability to turn off some focus on precise events
 # [0.3.5] 2021-02-18